fix: DIY-97 widen ownedStuffies bitmask past uint8_t's 8-item cap
Widens ownedStuffies to uint16_t following the same pattern as
DIY-90's ownedRoomThemes widening — equippedStuffy/equippedStuffyRight
stay uint8_t since they're single indices, not bitmasks. Adds a
static_assert(STUFFY_COUNT <= 16, ...) guard matching the one
ACCESSORY_COUNT/GLASSES_COUNT already have, closing the gap that let
this go unguarded in the first place.

Also fixes a latent JSON round-trip bug: fromJson()'s legacy-migration
check gated on doc["stuffies"].is<uint8_t>(), which would silently
fail once the bitmask value exceeds 255 (9+ owned stuffies), losing
real ownership data on load.

+++ apps/cyd-clock/src/ConfigManager.cpp
@@ -52,8 +52,11 @@ void ConfigManager::fromJson(JsonDocument& doc) {
     }
     _config.equippedBlanketColor = doc["blanketEquipped"] | _config.equippedBlanketColor;
     {
-        uint8_t stuffies;
-        if (doc["stuffies"].is<uint8_t>()) stuffies = doc["stuffies"];
+        // .is<uint16_t>(), not uint8_t: once 9+ stuffies are owned the bitmask value exceeds
+        // 255, and an is<uint8_t>() check would silently fail and fall through to the
+        // legacy/default branches below, losing real ownership data (DIY-97).
+        uint16_t stuffies;
+        if (doc["stuffies"].is<uint16_t>()) stuffies = doc["stuffies"];
         else if (doc["teddy"] | false) stuffies = 1;  // migrate legacy single-teddy flag
         else stuffies = _config.ownedStuffies;
         _config.ownedStuffies = stuffies;
